Phase 1: Extract results_writer.py from cli.py
Moves the CSV results-summary writer (create_results_file) into its own
module. Drops the lazy 'from rushti.utils import ...' inside the
function body in favor of a module-level import.

cli.py imports create_results_file at module scope so exit_rushti's
existing call site keeps working unchanged.

Also drops the now-unused 'csv' import from cli.py.
